% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{derivative}
\alias{derivative}
\title{calculates the derivative of a numeric vector}
\usage{
derivative(SERIES)
}
\arguments{
\item{SERIES}{a numeric vector}
}
\value{
numeric
}
\description{
calculates the derivative of a numeric vector
}
