% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{ema}
\alias{ema}
\title{calculates the exponential moving average of a numeric vector}
\usage{
ema(DAYS, SERIES)
}
\arguments{
\item{DAYS}{the number of days used to calculate the moving average}

\item{SERIES}{a numeric vector}
}
\value{
numeric
}
\description{
calculates the exponential moving average of a numeric vector
}
