% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_yfinance_file.R
\name{import_yfinance_file}
\alias{import_yfinance_file}
\title{imports a csv with historical data downloaded from Yahoo! Finance}
\usage{
import_yfinance_file(FILE)
}
\arguments{
\item{FILE}{character: a file path}
}
\value{
dataframe
}
\description{
imports a csv with historical data downloaded from Yahoo! Finance
}
