% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macd.R
\name{macd}
\alias{macd}
\title{calculates the moving average convergence/divergence (MACD) of a numeric vector}
\usage{
macd(FAST = 12, SLOW = 26, SIGNAL = 9, SERIES)
}
\arguments{
\item{FAST}{numeric: period of days for the fast series}

\item{SLOW}{numeric: period of days for the slow series}

\item{SIGNAL}{numeric: period of days for the signal}

\item{SERIES}{numeric vector of daily prices}
}
\value{
numeric
}
\description{
calculates the moving average convergence/divergence (MACD) of a numeric vector
}
