% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{calculates the moving average of a numeric vector}
\usage{
moving_average(DAYS, SERIES)
}
\arguments{
\item{DAYS}{the number of days used to calculate the moving average}

\item{SERIES}{a numeric vector}
}
\value{
numeric
}
\description{
calculates the moving average of a numeric vector
}
