% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsi.R
\name{rsi}
\alias{rsi}
\title{calculates the relative strength index (RSI) of a numeric vector}
\usage{
rsi(DAYS, CHANGE)
}
\arguments{
\item{DAYS}{the number of days used to calculate the RSI}

\item{CHANGE}{a numeric vector of daily price changes}
}
\value{
numeric
}
\description{
calculates the relative strength index (RSI) of a numeric vector
}
